add multi tenant features

+++ backend/src/services/product/src/models/product.model.ts
@@ -1,4 +1,9 @@
-import { generatePartialUpdate, IPaginationRequest, queryDb } from "shared";
+import {
+  generatePartialUpdate,
+  getTenantContext,
+  IPaginationRequest,
+  queryDb,
+} from "shared";
 import { PoolClient } from "pg";
 import {
   IProduct,
@@ -7,20 +12,19 @@ import {
 } from "../interfaces/product.interface";
 
 export async function createProduct(
-  { categoryId, name, slug, description, image = null }: IProduct,
+  { categoryId, name, slug, description, brand, image = null }: IProduct,
   options?: {
     client?: PoolClient;
   }
 ) {
-  const query = `INSERT INTO product(category_id,name,description,product_image,slug) VALUES($1,$2,$3,$4,$5) RETURNING *`;
+  const { tenantId } = getTenantContext();
+  const query = `INSERT INTO product(category_id,name,description,brand,product_image,slug,tenant_id) VALUES($1,$2,$3,$4,$5,$6,$7) RETURNING *`;
   const results = await queryDb(
     query,
-    [categoryId, name, description, image, slug],
+    [categoryId, name, description, brand, image, slug, tenantId],
     options
   );
-  const row = results?.rows?.[0];
-  if (!row) return null;
-  return row;
+  return results?.rows?.[0];
 }
 
 export async function createProductItem(
@@ -37,96 +41,50 @@ export async function createProductItem(
     client?: PoolClient;
   }
 ) {
+  const { tenantId } = getTenantContext();
   const query =
-    "INSERT INTO product_item(product_id,sku,qty_in_stock,product_image,price) VALUES($1,$2,$3,$4,$5) RETURNING id";
+    "INSERT INTO product_item(product_id,sku,qty_in_stock,product_image,price,tenant_id) VALUES($1,$2,$3,$4,$5,$6) RETURNING id";
   const results = await queryDb(
     query,
-    [productId, sku, qty_in_stock, product_image, price],
+    [productId, sku, qty_in_stock, product_image, price, tenantId],
     options
   );
-  const row = results?.rows[0];
-  if (!row) return null;
-  return row;
+  return results?.rows[0];
 }
 
 export async function getAllProducts({
   page = 1,
   limit = 20,
 }: IPaginationRequest) {
+  const { tenantId } = getTenantContext();
+
   const offset = (page - 1) * limit;
 
   const countResult = await queryDb(
-    `SELECT COUNT(*) AS total FROM product`,
-    []
+    `SELECT COUNT(*) AS total FROM product WHERE tenant_id=$1`,
+    [tenantId]
   );
   const total = Number(countResult?.rows?.[0]?.total || 0);
 
   const query = `
-  WITH paginated_products AS (
-    SELECT 
-    p.id, 
-    p.category_id, 
-    p.name, 
-    p.description, 
-    p.product_image
-    FROM product AS p 
-    ORDER BY p.id 
-    LIMIT $1 OFFSET $2
-),
-product_items_with_config AS (
-    SELECT 
-        pi.id AS product_item_id,
-        pi.product_id,
-        pi.sku,
-        pi.price,
-        pi.qty_in_stock,
-        pi.product_image,
-        COALESCE(json_agg(
-            jsonb_build_object(
-                'id', vo.id,
-                'value', vo.value,
-                'variation_id', vo.variation_id
-            )
-        ) FILTER (WHERE vo.id IS NOT NULL), '[]') AS configurations
-    FROM product_item  AS pi
-    LEFT JOIN product_configuration AS pc ON pc.product_item_id = pi.id
-    LEFT JOIN variation_option AS vo ON vo.id = pc.variation_option_id
-    GROUP BY pi.id
-),
-items_agg AS (
-    SELECT 
-     pi_config.product_id,
-    json_agg(
-            jsonb_build_object(
-                'id', pi_config.product_item_id,
-                'sku', pi_config.sku,
-                'price', pi_config.price,
-                'qty_in_stock', pi_config.qty_in_stock,
-                'product_image', pi_config.product_image,
-                'configuration', pi_config.configurations
-            )
-        ) AS items
-     FROM product_items_with_config  AS pi_config
-     GROUP BY pi_config.product_id
-)
-
 SELECT 
-    pp.id,
-    json_build_object(
-        'id', c.id,
-        'category_name', c.category_name
-    ) AS category,
-    pp.name,
-    pp.description,
-    pp.product_image,
-    COALESCE(i_agg.items, '[]') AS items
-FROM paginated_products AS pp 
-INNER JOIN product_category c ON c.id = pp.category_id
-LEFT JOIN items_agg as i_agg ON i_agg.product_id = pp.id
-ORDER BY pp.id;
+  p.id, 
+  p.name,
+  p.description,
+  p.product_image,
+  json_build_object('id', c.id, 'category_name', c.category_name) AS category,
+  COUNT(pi.id) AS item_count
+FROM product AS p
+JOIN product_category AS c
+  ON c.id = p.category_id
+LEFT JOIN product_item AS pi
+  ON pi.product_id = p.id
+WHERE p.tenant_id = $3
+GROUP BY p.id, c.id
+LIMIT $1 OFFSET $2;
   `;
 
-  const results = await queryDb(query, [limit, offset]);
+  const results = await queryDb(query, [limit, offset, tenantId]);
   const data = results.rows;
 
   const totalPages = Math.ceil(total / limit);
@@ -148,12 +106,11 @@ export async function getProductById(
     client: PoolClient;
   }
 ) {
-  const query = `SELECT id, category_id, name, description, product_image FROM product WHERE id = $1`;
+  const { tenantId } = getTenantContext();
+  const query = `SELECT id, category_id, name, description, brand, product_image FROM product WHERE id = $1 AND tenant_id=$2`;
 
-  const results = await queryDb(query, [id], options);
-  const row = results?.rows[0];
-  if (!row) return null;
-  return row;
+  const results = await queryDb(query, [id, tenantId], options);
+  return results?.rows[0];
 }
 
 export async function getProductWithCategoryById(
@@ -162,6 +119,7 @@ export async function getProductWithCategoryById(
     client?: PoolClient;
   }
 ) {
+  const { tenantId } = getTenantContext();
   const query = `
   SELECT 
   p.id,
@@ -172,25 +130,25 @@ export async function getProductWithCategoryById(
   p.name,
   p.slug,
   p.description,
+  p.brand,
   p.product_image
   FROM product AS p
   INNER JOIN product_category AS c 
     ON p.category_id = c.id
-  WHERE p.id = $1
+  WHERE p.id = $1 AND tenant_id=$2
   `;
 
-  const results = await queryDb(query, [id], options);
-  const row = results?.rows[0];
-  if (!row) return null;
-  return row;
+  const results = await queryDb(query, [id, tenantId], options);
+  return results?.rows[0];
 }
 
 export async function getProductItemsByIds(
   itemIds: IProductItem["id"][],
   options?: { client?: PoolClient }
 ) {
-  const query = "SELECT id FROM product_item WHERE id=ANY($1)";
-  const results = await queryDb(query, [itemIds], options);
+  const { tenantId } = getTenantContext();
+  const query = "SELECT id FROM product_item WHERE id=ANY($1) AND tenant_id=$2";
+  const results = await queryDb(query, [itemIds, tenantId], options);
   return results.rows;
 }
 
@@ -199,8 +157,10 @@ export async function checkSkuAlreadyExists(
   selfId: number,
   options?: { client?: PoolClient }
 ) {
-  const query = "SELECT id FROM product_item WHERE sku = $1 AND id != $2";
-  const results = await queryDb(query, [sku, selfId], options);
+  const { tenantId } = getTenantContext();
+  const query =
+    "SELECT id FROM product_item WHERE sku = $1 AND id != $2 AND tenant_id=$3";
+  const results = await queryDb(query, [sku, selfId, tenantId], options);
   return results.rows?.length > 0;
 }
 
@@ -210,9 +170,12 @@ export async function getProductItemById(
     client?: PoolClient;
   }
 ) {
-  const query = `SELECT id, product_id, sku, qty_in_stock, product_image, price FROM product_item WHERE id = $1`;
+  const { tenantId } = getTenantContext();
+  const query = `SELECT id, product_id, sku, qty_in_stock, product_image, price 
+  FROM product_item 
+  WHERE id = $1 AND tenant_id=$2`;
 
-  const results = await queryDb(query, [id], options);
+  const results = await queryDb(query, [id, tenantId], options);
   const row = results?.rows[0];
   if (!row) return null;
   return row;
@@ -222,8 +185,9 @@ export async function getProductItemsByProductId(
   productId: IProduct["id"],
   options?: { client?: PoolClient }
 ) {
-  const query = `SELECT * FROM product_item WHERE product_id = $1 ORDER BY id`;
-  const results = await queryDb(query, [productId], options);
+  const { tenantId } = getTenantContext();
+  const query = `SELECT * FROM product_item WHERE product_id = $1 AND tenant_id=$2 ORDER BY id`;
+  const results = await queryDb(query, [productId, tenantId], options);
   return results?.rows;
 }
 
@@ -231,8 +195,9 @@ export async function getProductItemBySku(
   sku: string,
   options?: { client?: PoolClient }
 ) {
-  const query = `SELECT * FROM product_item WHERE sku = $1`;
-  const results = await queryDb(query, [sku], options);
+  const { tenantId } = getTenantContext();
+  const query = `SELECT * FROM product_item WHERE sku = $1 AND tenant_id=$2`;
+  const results = await queryDb(query, [sku, tenantId], options);
   return results?.rows[0];
 }
 
@@ -242,19 +207,20 @@ export async function getProductConfigurationById(
     client?: PoolClient;
   }
 ) {
-  const query = `SELECT * FROM product_configuration WHERE id = $1 `;
-  const results = await queryDb(query, [id], options);
-  const row = results?.rows[0];
-  if (!row) return null;
-  return row;
+  const { tenantId } = getTenantContext();
+  const query = `SELECT * FROM product_configuration WHERE id = $1 AND tenant_id=$2`;
+  const results = await queryDb(query, [id, tenantId], options);
+  return results?.rows[0];
 }
 
 export async function getConfigurationsByIds(
   configIds: IProductConfiguration["id"][],
   options?: { client?: PoolClient }
 ) {
-  const query = "SELECT id FROM product_configuration WHERE id=ANY($1)";
-  const results = await queryDb(query, [configIds], options);
+  const { tenantId } = getTenantContext();
+  const query =
+    "SELECT id FROM product_configuration WHERE tenant_id=$1 AND id=ANY($2)";
+  const results = await queryDb(query, [tenantId, configIds], options);
   return results.rows;
 }
 
@@ -264,6 +230,7 @@ export async function getConfigurationsByProductItemIds(
     client?: PoolClient;
   }
 ) {
+  const { tenantId } = getTenantContext();
   const query = `
   SELECT 
   pc.product_item_id,
@@ -280,34 +247,42 @@ export async function getConfigurationsByProductItemIds(
   FROM product_configuration AS pc 
   JOIN variation_option AS vo
     ON vo.id = pc.variation_option_id
-  WHERE pc.product_item_id = ANY($1)
+  WHERE pc.tenant_id =$1 AND  pc.product_item_id = ANY($2)
   GROUP BY pc.product_item_id
   `;
-  const results = await queryDb(query, [productItemIds], options);
+  const results = await queryDb(query, [tenantId, productItemIds], options);
   return results?.rows;
 }
 
 export async function createProductConfiguration(
   productItemId: number,
+  variationId: number,
   variationOptionId: number,
   options?: {
     client?: PoolClient;
   }
 ) {
+  const { tenantId } = getTenantContext();
   const query =
-    "INSERT INTO product_configuration(product_item_id,variation_option_id) VALUES($1,$2)";
-  await queryDb(query, [productItemId, variationOptionId], options);
+    "INSERT INTO product_configuration(product_item_id,variation_id,variation_option_id,tenant_id) VALUES($1,$2,$3,$4)";
+  await queryDb(
+    query,
+    [productItemId, variationId, variationOptionId, tenantId],
+    options
+  );
 }
 
 export async function updateProduct(
   productId: number,
-  { categoryId, description, image, name }: Partial<IProduct>,
+  { categoryId, description, image, brand, name }: Partial<IProduct>,
   options?: { client?: PoolClient }
 ) {
+  const { tenantId } = getTenantContext();
   const partialUpdate = generatePartialUpdate({
     category_id: categoryId,
     name,
     description,
+    brand,
     product_image: image,
   });
   if (!partialUpdate) return;
@@ -315,10 +290,12 @@ export async function updateProduct(
   const query = [
     "UPDATE product SET",
     partialUpdate.setString,
-    `WHERE id=$${partialUpdate.values.length + 1} RETURNING id`,
+    `WHERE id=$${partialUpdate.values.length + 1} AND tenant_id=$${
+      partialUpdate.values.length + 2
+    } RETURNING id`,
   ].join(" ");
 
-  await queryDb(query, [...partialUpdate.values, productId], options);
+  await queryDb(query, [...partialUpdate.values, productId, tenantId], options);
 }
 
 export async function updateProductItem(
@@ -328,6 +305,7 @@ export async function updateProductItem(
     client?: PoolClient;
   }
 ) {
+  const { tenantId } = getTenantContext();
   const partialUpdate = generatePartialUpdate({
     product_image,
     qty_in_stock,
@@ -339,10 +317,12 @@ export async function updateProductItem(
   const query = [
     "UPDATE product_item SET",
     partialUpdate.setString,
-    `WHERE id=$${partialUpdate.values.length + 1} RETURNING id`,
+    `WHERE id=$${partialUpdate.values.length + 1} AND tenant_id=$${
+      partialUpdate.values.length + 2
+    } RETURNING id`,
   ].join(" ");
 
-  await queryDb(query, [...partialUpdate.values, id], options);
+  await queryDb(query, [...partialUpdate.values, id, tenantId], options);
 }
 
 export async function updateProductConfiguration(
@@ -353,14 +333,20 @@ export async function updateProductConfiguration(
     client?: PoolClient;
   }
 ) {
+  const { tenantId } = getTenantContext();
   const query =
-    "UPDATE product_configuration SET variation_option_id=$1 WHERE product_item_id=$2 AND id=$3";
-  await queryDb(query, [variationOptionId, productItemId, id], options);
+    "UPDATE product_configuration SET variation_option_id=$1 WHERE product_item_id=$2 AND id=$3 AND tenant_id=$4";
+  await queryDb(
+    query,
+    [variationOptionId, productItemId, id, tenantId],
+    options
+  );
 }
 
 export async function deleteProduct(id: number) {
-  const query = "DELETE FROM product WHERE id=$1 RETURNING id";
-  const results = await queryDb(query, [id]);
+  const { tenantId } = getTenantContext();
+  const query = "DELETE FROM product WHERE id=$1 AND tenant_id=$2 RETURNING id";
+  const results = await queryDb(query, [id, tenantId]);
   const row = results?.rows[0];
   if (!row) return null;
   return row;
